requirements.txt
PyJWT>=1.7.1
cachetools>=4.0.0
certifi>=2020.4.5.1
cffi>=1.14.0
chardet>=3.0.4
cryptography>=2.9
cycler>=0.10.0
google-api-core>=1.16.0
google-api-python-client>=1.8.0
google-auth>=1.13.1
google-auth-httplib2>=0.0.3
google-cloud-pubsub>=1.4.2
google-cloud-firestore>=1.5.0
google-cloud-iot >= 1.0.0
googleapis-common-protos>=1.51.0
grpc-google-iam-v1>=0.12.3
grpcio>=1.28.1
httplib2>=0.17.1
idna>=2.9
paho-mqtt>=1.5.0
protobuf>=3.11.3
psycopg2>=2.8.5
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pycparser>=2.20
pyparsing>=2.4.7
python-dateutil>=2.8.1
pytz>=2019.3
requests>=2.23.0
rsa>=4.0
six>=1.14.0
uritemplate>=3.0.1
urllib3>=1.25.8